+++ ic-utils/src/interfaces.rs
@@ -1,7 +1,9 @@
+pub mod bitcoin_canister;
 pub mod http_request;
 pub mod management_canister;
 pub mod wallet;
 
+pub use bitcoin_canister::BitcoinCanister;
 pub use http_request::HttpRequestCanister;
 pub use management_canister::ManagementCanister;
 pub use wallet::WalletCanister;